Refactor SeedAddress into separate component
